+++ ts/packages/knowPro/src/interfaces.ts
@@ -8,13 +8,6 @@ export interface IKnowledgeSource {
     getKnowledge(): kpLib.KnowledgeResponse | undefined;
 }
 
-export interface IReadonlyCollection<T, TOrdinal> extends Iterable<T> {
-    readonly length: number;
-    get(ordinal: TOrdinal): T | undefined;
-    getMultiple(ordinals: TOrdinal[]): (T | undefined)[];
-    getAll(): T[];
-}
-
 export type MessageOrdinal = number;
 
 /**
@@ -305,3 +298,18 @@ export type ListIndexingResult = {
     numberCompleted: number;
     error?: string | undefined;
 };
+
+//---------------------
+// Collections
+//---------------------
+export interface IReadonlyCollection<T, TOrdinal> extends Iterable<T> {
+    readonly length: number;
+    get(ordinal: TOrdinal): T | undefined;
+    getMultiple(ordinals: TOrdinal[]): (T | undefined)[];
+    getAll(): T[];
+}
+
+export interface ICollection<T, TOrdinal>
+    extends IReadonlyCollection<T, TOrdinal> {
+    push(...items: T[]): void;
+}
