+++ ts/packages/memory/storage/src/sqlite/index.ts
@@ -1,4 +1,4 @@
 // Copyright (c) Microsoft Corporation.
 // Licensed under the MIT License.
 
-export * from "./sqliteCollections.js";
+export * from "./sqliteCollection.js";
